handle libvirtd restart on CentOS 7
libvirtd on CentOS 7 will inject its iptables configuration whenever libvirtd
is restarted. That means VMs with routed IPs are not accessible anymore after
"systemctl restart libvirtd".

This is especially problematic as RPM updates can trigger a restart of
libvirtd.

Therefore we listen to the "plugged" event which is triggered whenever a new
VM is started. Conveniently this event is also triggered after a daemon
restart if a VM is already running.

However that means we should check if the iptables rules are already present
in the iptables chain as the "plugged" event is triggered for each VM and
I don't want to add the same rules multiple times if there are multiple VMs
with routed IPs inside the same libvirt network.

The "started" event is only used to configure the actual route ("ip route add").
This makes the code simpler as I don't have to check if the route is already
present. libvirtd only overrides iptables rules but not the routes.